requirements.txt
Flask==0.11.1
Flask-PyMongo==0.4.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.6.0
Jinja2==2.8
jsonschema==3.0.1
numpy==1.16.2
Pillow==5.4.1
plotly==3.7.0
psutil==5.6.1
pymongo==3.7.2
requests==2.21.0
requests-toolbelt==0.9.1
Unirest==1.1.7
urllib3==1.24.1
